Changing names in the build service. - The names of methods, structures and interfaces in the service for creating block structures have been changed. - The names of methods, structures and interfaces in the configuration template structure creation service have been changed.
