+++ internal/service/builder/builder.go
@@ -1,7 +1,7 @@
 package builder
 
 var (
-	structBuilder StructureBuilder = &Structure{}
-	changes       ChangesStructure = &Changes{}
-	pathProject   string
+	blockBuilder BlockBuilder     = &Block{}
+	changes      ChangesStructure = &Changes{}
+	pathProject  string
 )
